+++ apps/block_scout_web/lib/block_scout_web/api_router.ex
@@ -326,8 +326,13 @@ defmodule BlockScoutWeb.ApiRouter do
     scope "/proxy" do
       scope "/noves-fi" do
         get("/transactions/:transaction_hash_param", V2.Proxy.NovesFiController, :transaction)
-        get("/transactions/:transaction_hash_param/describe", V2.Proxy.NovesFiController, :describe_transaction)
-        get("/addresses/:address_hash_param/transactions", V2.Proxy.NovesFiController, :address_transactions)
+
+        get(
+          "/addresses/:address_hash_param/transaction-descriptions",
+          V2.Proxy.NovesFiController,
+          :address_transactions
+        )
+
         get("/transactions", V2.Proxy.NovesFiController, :describe_transactions)
       end