+++ apps/block_scout_web/lib/block_scout_web/controllers/api/v2/proxy/noves_fi_controller.ex
@@ -26,26 +26,7 @@ defmodule BlockScoutWeb.API.V2.Proxy.NovesFiController do
   end
 
   @doc """
-    Function to handle GET requests to `/api/v2/proxy/noves-fi/transactions/:transaction_hash_param/describe` endpoint.
-  """
-  @spec describe_transaction(Plug.Conn.t(), map()) :: Plug.Conn.t() | {atom(), any()}
-  def describe_transaction(conn, %{"transaction_hash_param" => transaction_hash_string} = params) do
-    with {:ok, _transaction, _transaction_hash} <-
-           TransactionController.validate_transaction(transaction_hash_string, params,
-             necessity_by_association: %{},
-             api?: true
-           ),
-         url = NovesFi.describe_tx_url(transaction_hash_string),
-         {response, status} <- NovesFi.noves_fi_api_request(url, conn),
-         {:is_empty_response, false} <- {:is_empty_response, is_nil(response)} do
-      conn
-      |> put_status(status)
-      |> json(response)
-    end
-  end
-
-  @doc """
-    Function to handle GET requests to `/api/v2/proxy/noves-fi/transactions/:transaction_hash_param/transactions` endpoint.
+    Function to handle GET requests to `/api/v2/proxy/noves-fi/transactions/:transaction_hash_param/transaction-descriptions` endpoint.
   """
   @spec address_transactions(Plug.Conn.t(), map()) :: Plug.Conn.t() | {atom(), any()}
   def address_transactions(conn, %{"address_hash_param" => address_hash_string} = params) do